Drawing the held items to the screen and making the attack button shoot projectiles
Using the same logic from PlaceItem(), the current inventory item is now placed so the player is holding it.
The attack button z now shoots the projectiles.
In the future new projectiles should be added such as a sword slash for the sword which will be stationary, then disappear. And a bomb explosion.