Added exploding bombs
Bomb when placed causes an explosion after some time.
Bombs are consumable, so when it is used it removes it from the inventory. Still need to implement so it only removes one bomb instead of all of them.
Added new fields to IItem, ToBeDeleted and Equipped to separate world items and inventory items.
When ToBeDeleted is true, an item is collected
When Equipped is false, an item is removed from the inventory.